+++ ui/src/components/General/ClipboardButton/utils/types.ts
@@ -1,0 +1,8 @@
+interface ClipboardButtonProps {
+	copyToClipboard: (text: string) => void;
+	copied: boolean;
+	text: string;
+	label?: boolean;
+}
+
+export type { ClipboardButtonProps };