+++ ui/src/components/General/InputComponent/utils/types.ts
@@ -1,0 +1,23 @@
+type NumberType = {
+	type: "number";
+	value: number;
+	onChange: (value: number | null) => void;
+};
+
+type TextType = {
+	type: "text";
+	value: string;
+	onChange: (value: React.ChangeEvent<HTMLInputElement>) => void;
+};
+
+type InputComponentPropsType = {
+	label: string;
+	precision?: number;
+	min?: number;
+	max?: number;
+	step?: number;
+	tooltip?: string;
+	placeholder?: string;
+} & (NumberType | TextType);
+
+export type { InputComponentPropsType };