+++ ui/src/components/General/Text/utils/constant.ts
@@ -1,0 +1,3 @@
+const LIST_PAGE_COMMON_PATH = "list";
+
+export { LIST_PAGE_COMMON_PATH };