+++ ui/src/components/General/Text/utils/types.ts
@@ -1,0 +1,6 @@
+interface TextProps {
+	text: string;
+	level?: 1 | 2 | 3 | 4 | 5;
+}
+
+export type { TextProps };