+++ ui/src/components/Layouts/Footer/tests/Footer.test.tsx
@@ -1,10 +1,10 @@
 import { render, screen } from "@testing-library/react";
-import { describe } from "vitest";
-import Footer from "..";
+import { describe, vi } from "vitest";
+import Footer from "components/Layouts/Footer";
 
 describe("Footer", () => {
 	test("render component without crash", () => {
-		render(<Footer />);
+		render(<Footer isDarkMode={false} handleThemeChange={vi.fn} />);
 
 		const imageElement = screen.getByRole("img");