+++ ui/src/components/Layouts/Header/utils/types.ts
@@ -1,0 +1,8 @@
+interface HeaderProps {
+	handleThemeChange: () => void;
+	isDarkMode: boolean;
+	collapsed: boolean;
+	handleMenuCollapse: () => void;
+}
+
+export type { HeaderProps };