+++ ui/src/components/Layouts/Menu/index.tsx
@@ -1,15 +1,13 @@
 import { Menu as AntdMenu, Layout } from "antd";
 import { To, useNavigate } from "react-router-dom";
 import { ITEMS } from "./utils/constant";
+import { MenuProps } from "./utils/types";
+import style from "./Menu.module.scss";
+import { classNames } from "lib/utils/helper";
 
 const { Sider } = Layout;
 
-interface MenuProps {
-	isDarkMode: boolean;
-	collapsed: boolean;
-}
-
-const Menu: React.FC<MenuProps> = ({ collapsed }) => {
+const Menu: React.FC<MenuProps> = ({ collapsed, isDarkMode }) => {
 	const navigate = useNavigate();
 
 	const onClick = (e: { key: To }) => {
@@ -19,7 +17,10 @@ const Menu: React.FC<MenuProps> = ({ collapsed }) => {
 	return (
 		<Sider trigger={null} collapsible collapsed={collapsed}>
 			<AntdMenu
-				style={{ borderInlineEnd: "none" }}
+				className={classNames(
+					style.menu,
+					isDarkMode ? style.menu__dark : undefined
+				)}
 				mode="inline"
 				onClick={onClick}
 				items={ITEMS}