+++ ui/src/components/RenderProps/Clipboard/utils/test.ts
@@ -1,0 +1,8 @@
+interface ClipboardProps {
+	text: string;
+	clipboardComponent: React.ElementType;
+	className?: string;
+	label?: string;
+}
+
+export type { ClipboardProps };