+++ ui/src/lib/utils/context/DarkModeProvider/utils/constant.ts
@@ -1,0 +1,3 @@
+const DARK_MODE_STORAGE_KEY = "darkMode";
+
+export { DARK_MODE_STORAGE_KEY };