+++ ui/src/pages/List/Blog/index.tsx
@@ -1,9 +1,7 @@
 import Resource from "components/General/ListItems/Resource/Resource";
-import ListSearchResults from "components/RenderProps/ListSearchResults/ListSearchResults";
+import ListSearchResults from "components/RenderProps/ListSearchResults";
 import useFetchList from "lib/utils/hooks/useFetchList";
-
-const URL = `./blog.json`;
-const QUERY_KEY = "blog";
+import { QUERY_KEY, URL } from "./utils/contant";
 
 const Blog: React.FC = () => {
 	const { data, isLoading, isError } = useFetchList(QUERY_KEY, URL);