+++ ui/src/pages/List/Book/utils/contant.ts
@@ -1,0 +1,4 @@
+const URL = `./book.json`;
+const QUERY_KEY = "book";
+
+export { URL, QUERY_KEY };