+++ ui/src/pages/List/Course/utils/contant.ts
@@ -1,0 +1,4 @@
+const URL = `./course.json`;
+const QUERY_KEY = "course";
+
+export { URL, QUERY_KEY };