+++ ui/src/pages/List/Movie/utils/contant.ts
@@ -1,0 +1,4 @@
+const URL = `./movie.json`;
+const QUERY_KEY = "movie";
+
+export { URL, QUERY_KEY };