+++ ui/src/pages/List/Tool/utils/contant.ts
@@ -1,0 +1,4 @@
+const URL = `./tool.json`;
+const QUERY_KEY = "tool";
+
+export { URL, QUERY_KEY };