+++ ui/src/pages/List/YouTube/utils/contant.ts
@@ -1,0 +1,4 @@
+const URL = `./youtube.json`;
+const QUERY_KEY = "youtube";
+
+export { URL, QUERY_KEY };